Some things
I changed project build settings, so preprocessing is optimized for the
BT-200. This should increase the size of the final build, but will
increase performance significantly.

Do not pay attention to the scene files yet, those are just for testing.
I am currently testing how to properly make the AR camera behave
independent of the object being tracked, and I will eventually look into
merging the Moverio Controller with the ARCamera so we can use sensors
etc as intended.

Priority right now is getting frame markers to work without the camera
changing the world center to the marker.