ARExampleInteractable
Added ARExampleInteractable, you need to assign it to the respective
ARObject under ARObjects > <name> > Tracked Objects: increase the list
by one and add your object to the end of the list.
Added the "level", but prefabs still need to be made!
Fixed the ARPosition null pointer references.
Adjusted camera position, still needs to be angled a bit on global-y.
Counters can be visible in the editor now.

TODO: Jumping, need to stop the agent before doing the jump and resume
it afterwards!