Added jump and stop functionality
Added jumping and stopping triggers, the player will follow the
direction the "target" GameObject is pointing at. So to make it go
right, point it to the right and to make it go left point it left.

Reorganized the folder structure, everything you need for level making
is under KrankenFabs > Level.