Fixes
Fixed some shadows.
Made the AR box a bit transparent.
Updated SjoerTestLevel, you can now start it and get to the end to grab
the coin. Use it as an example, as I will continue using this test setup
to add more elements that the designers can work with.
Added the ToggleActive action, it is used for buttons/switches to turn
GameObjects on or off.
Added a dirty hack to stay at Z = 0.0f;
Added variable jump power to the JumpTrigger.
Added a TurnTrigger.